more arrangement and add redis client example
