+++ jumpscale/clients/redis/__init__.py
@@ -1,0 +1,6 @@
+from jumpscale.core.base import StoredFactory
+
+from .redis import Redis
+
+
+factory = StoredFactory(Redis)
