+++ jumpscale/core/base/__init__.py
@@ -1,2 +1,3 @@
 from .meta import Base
-from .factory import StoredFactory
+from .fields import ValidationError
+from .factory import Factory, StoredFactory, DuplicateError
