+++ jumpscale/core/base/meta.py
@@ -1,73 +1,8 @@
 from collections import namedtuple
 from types import SimpleNamespace
 
-
-class DuplicateError(Exception):
-    pass
-
-
-class ValidationError(Exception):
-    pass
-
-
-class Field:
-    def __init__(self, **kwargs):
-        self.default = None
-        self.required = False
-        self.kwargs = kwargs
-
-    def validate(self, value):
-        pass
-
-
-class Factory:
-    def __init__(self, type_):
-        self.type = type_
-        self.__count = 0
-        super(Factory, self).__init__()
-
-    def new(self, name, *args, **kwargs):
-        if not name.isidentifier():
-            raise ValueError("%s is not a valid identifier" % name)
-
-        try:
-            self.find(name)
-        except AttributeError:
-            instance = self.type(*args, **kwargs)
-            setattr(self, name, instance)
-
-            self.count += 1
-            return instance
-
-        raise DuplicateError
-
-    def find(self, name):
-        instance = getattr(self, name)
-        if not isinstance(instance, self.type):
-            raise ValueError("%s is an internal attribute" % name)
-        return instance
-
-    def get(self, name, *args, **kwargs):
-        try:
-            return self.find(name)
-        except AttributeError:
-            return self.new(name, *args, **kwargs)
-
-    def delete(self, name):
-        self.count -= 1
-        delattr(self, name)
-
-    def updated(self, new_count):
-        pass
-
-    def get_count(self):
-        return self.__count
-
-    def set_count(self, new_count):
-        self.__count = new_count
-        self.updated(new_count)
-
-    count = property(get_count, set_count)
+from .factory import Factory, DuplicateError
+from .fields import Field, ValidationError
 
 
 def get_field_property(name, field):
